pyproject.toml
[project]
name = "personal-agent-os"
version = "0.1.0"
description = "Personal AI agent OS for autonomous task automation"
authors = [{ name = "Masahiro Ishizuka", email = "ishimasa01@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic (>=0.75.0,<0.76.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-api-python-client (>=2.100.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<1.0.0)",
    "google-auth-oauthlib (>=1.2.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
